refact: queryKey constants 정리 (#4)
