+++ src/constants/queryKeyConstants.js
@@ -1,0 +1,9 @@
+
+export const PlACES_QUERY_KEY = {
+  PlACES: ["places"],
+  PLACE_PLACE_ID: (id)=> ["place", id],
+}
+
+export const COMMENT_QUERY_KEY = {
+  COMMENT: ["comment"],
+}
